Separating out the URL
Removing the URL info from the initial call to speed up load time of recipe tiles.

Creating a separate function that when called  returns the URL.

With this approach 1 call is made to load all the recipes initially and a second call is made at the request of the link to the original recipe. This is compares to 1 call to get all the recipes and the n (number of recipes) calls for the extra info which was just the source URL